fix: [stix2 import] Fixed email object import (screenshot & eml attributes)
- Same comments as previous commit for export
- Also moved parsing functions in subject into the
  main script to avoid importing python libraries
  in the dictionaries script
