+++ app/files/scripts/stix2/stix2misp.py
@@ -16,7 +16,7 @@
 #    You should have received a copy of the GNU Affero General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import sys, json, os, time, uuid
+import sys, json, os, time, uuid, io
 import stix2
 from pymisp import MISPEvent, MISPObject, __path__
 from stix2misp_mapping import *
@@ -90,7 +90,7 @@ class StixParser():
     def load_mapping(self):
         self.objects_mapping = {'asn': {'observable': observable_asn, 'pattern': pattern_asn},
                                 'domain-ip': {'observable': observable_domain_ip, 'pattern': pattern_domain_ip},
-                                'email': {'observable': observable_email, 'pattern': pattern_email},
+                                'email': {'observable': self.observable_email, 'pattern': self.pattern_email},
                                 'file': {'observable': observable_file, 'pattern': pattern_file},
                                 'ip-port': {'observable': observable_ip_port, 'pattern': pattern_ip_port},
                                 'network-socket': {'observable': observable_socket, 'pattern': pattern_socket},
@@ -269,6 +269,85 @@ class StixParser():
             self.misp_event.add_attribute(**attribute)
         except:
             pass
+
+    @staticmethod
+    def observable_email(observable):
+        attributes = []
+        addresses = {}
+        files = {}
+        for o_key, o_dict in observable.items():
+            part_type = o_dict._type
+            if part_type == 'email-addr':
+                addresses[o_key] = o_dict.get('value')
+            elif part_type == 'file':
+                files[o_key] = o_dict.get('name')
+            else:
+                message = dict(o_dict)
+        attributes.append({'type': 'email-src', 'object_relation': 'from',
+                           'value': addresses[message.pop('from_ref')], 'to_ids': False})
+        for ref in ('to_refs', 'cc_refs'):
+            if ref in message:
+                for item in message.pop(ref):
+                    mapping = email_mapping[ref]
+                    attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
+                                       'value': addresses[item], 'to_ids': False})
+        if 'body_multipart' in message:
+            for f in message.pop('body_multipart'):
+                attributes.append({'type': 'email-attachment', 'object_relation': 'attachment',
+                                   'value': files[f.get('body_raw_ref')], 'to_ids': False})
+        for m_key, m_value in message.items():
+            if m_key == 'additional_header_fields':
+                for field_key, field_value in m_value.items():
+                    mapping = email_mapping[field_key]
+                    if field_key == 'Reply-To':
+                        for rt in field_value:
+                            attributes.append({'type': mapping['type'],
+                                               'object_relation': mapping['relation'],
+                                               'value': rt, 'to_ids': False})
+                    else:
+                        attributes.append({'type': mapping['type'],
+                                           'object_relation': mapping['relation'],
+                                           'value': field_value, 'to_ids': False})
+            else:
+                try:
+                    mapping = email_mapping[m_key]
+                    attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
+                                       'value': m_value, 'to_ids': False})
+                except:
+                    if m_key.startswith("x_misp_attachment_"):
+                        attribute_type, relation = m_key.split("x_misp_")[1].split("_")
+                        attributes.append({'type': attribute_type, 'object_relation': relation, 'to_ids': False,
+                                           'value': m_value['value'], 'data': io.BytesIO(m_value['data'].encode())})
+                    elif "x_misp_" in m_key:
+                        attribute_type, relation = m_key.split("x_misp_")[1].split("_")
+                        attributes.append({'type': attribute_type, 'object_relation': relation,
+                                           'value': m_value, 'to_ids': False})
+        return attributes
+
+    @staticmethod
+    def pattern_email(pattern):
+        attributes = []
+        attachments = defaultdict(dict)
+        for p in pattern:
+            p_type, p_value = p.split(' = ')
+            try:
+                mapping = email_mapping[p_type]
+                attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
+                                   'value': p_value[1:-1], 'to_ids': True})
+            except KeyError:
+                if p_type.startswith("email-message:'x_misp_attachment_"):
+                    relation, field = p_type.split('.')
+                    relation = relation.split(':')[1][1:-1]
+                    attachments[relation][field] = p_value[1:-1]
+                elif "x_misp_" in p_type:
+                    attribute_type, relation = p_type.split("x_misp_")[1][:-1].split("_")
+                    attributes.append({'type': attribute_type, 'object_relation': relation,
+                                       'value': p_value[1:-1], 'to_ids': True})
+        for a_key, a_dict in attachments.items():
+            _, _, attribute_type, relation = a_key.split('_')
+            attributes.append({'type': attribute_type, 'object_relation': relation, 'to_ids': True,
+                               'value': a_dict['value'], 'data': io.BytesIO(a_dict['data'].encode())})
+        return attributes
 
     def observable_pe(self, observable):
         extension = observable['0']['extensions']['windows-pebinary-ext']
